refactor(viewer): split CellBuilderPanel into section components
CellBuilderPanel.tsx was 2,189 lines: six tab bodies, a header, a footer,
two inline hooks and four helper components in one file. Each is now its
own component under components/viewer/cellbuilder/panel/, and the panel
file (165 lines) is the composition — which section renders where, plus
the small local UI state (which tab, which popover) they share.

Sections select what they render from the store themselves (the panel
already re-rendered them all on any store change, so this is the same
subscription); only genuinely local state is passed as props: the two
add-mode popovers to BuildTab, the local-model browser to PanelHeader,
and an `active` flag to the two catalog tabs.

New files: chrome (shared CSS tokens, Section, isReadOnly,
describeToolState), IconOverlaySection, CompileLogSection, SystemsTab,
ConnectionAdder, useLocalModelBrowser, useSheetDrag, PanelHeader,
PanelTabBar, BuildTab, CompileSettingsSection, GroupsSection,
EquipmentTab, SystemsCatalogTab, ViewTab, ToolsTab, PanelFooter.

Behaviour-preserving: JSX is moved verbatim (no user-visible string
changed), every tab body stays MOUNTED-but-hidden as before so
SystemsTab's auto-highlight effect keeps tracking a loaded result, and
the sheet-drag hook takes its dismiss action as a callback instead of
reaching into the store. Largest resulting file is 395 lines.
